Extract filename/title/lookup formatters into formatters module
Move the twelve filename/title/lookup formatters out of `utils.py`
into a new `formatters` module:

- AOS_CCDS and AOS_WORKER_MAPPING (the (depth, ccd) flattening table)
- FakeExposureRecord — the dataclass stand-in used when there is no
  real DimensionRecord available for a filename
- expRecordToUploadFilename — RubinTV upload filename builder
- getRubinTvInstrumentName — instrument name normalisation
- getFilterColorName — physical filter → RubinTV cell colour lookup
- mapAosWorkerNumber — worker index → (depth, ccd) lookup
- getPodWorkerNumber — env-driven worker number
- makePlotFile / makePlotFileFromRecord — plot file path builders
- makeWitnessDetectorTitle / makeFocalPlaneTitle — plot title builders

These are all small string-/path-/lookup-table builders that have no
runtime dependency on anything else in `utils.py`. They form the
natural second carve-out of the larger split, after `predicates`.

All callers across the package, scripts, and tests have been updated
to import from `formatters` instead of `utils`. The previously-unused
imports of `itertools`, `pathlib.Path`, `dayObsIntToString`,
`PREFIXES` and `Camera` have been removed from `utils.py`. As per the
"this is an application, not a library" convention, no
compatibility shims are added: the moved symbols are deleted outright
from `utils.py` and every caller is updated in the same change.

The TYPE_CHECKING import of `LocationConfig` in the new module still
points at `.utils` for now, since `LocationConfig` itself has not
been moved yet. It will be re-pointed when the locationConfig module
is extracted in a later commit.

Verified:
- mypy and flake8 clean on both changed package files and on the
  scripts directory.
- The full unit test suite (185 tests) passes — the same predicates
  and AOS_CCDS-using tests continue to pass against their new homes.
- Every dependent in-package module imports cleanly.

Written by Claude Code 4.6 Opus (high-effort).
