+++ python/lsst/rubintv/production/highLevelTools.py
@@ -48,8 +48,9 @@ from lsst.utils.iteration import sequence_to_string
 
 from .channels import CHANNELS, PREFIXES
 from .consdbUtils import CCD_VISIT_MAPPING, ConsDBPopulator, changeType
+from .formatters import FakeExposureRecord, expRecordToUploadFilename
 from .uploaders import Uploader
-from .utils import FakeExposureRecord, LocationConfig, expRecordToUploadFilename
+from .utils import LocationConfig
 
 HAS_EFD_CLIENT = True
 try:
