+++ scripts/LSSTCam/runNightlyWorker.py
@@ -22,14 +22,11 @@
 import sys
 
 from lsst.daf.butler import Butler
+from lsst.rubintv.production.formatters import getPodWorkerNumber
 from lsst.rubintv.production.pipelineRunning import SingleCorePipelineRunner
 from lsst.rubintv.production.podDefinition import PodDetails, PodFlavor
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import (
-    getAutomaticLocationConfig,
-    getPodWorkerNumber,
-    setupSentry,
-)
+from lsst.rubintv.production.utils import getAutomaticLocationConfig, setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
