+++ scripts/LSSTComCamSim/runAosWorker.py
@@ -20,10 +20,11 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from lsst.daf.butler import Butler
+from lsst.rubintv.production.formatters import getPodWorkerNumber
 from lsst.rubintv.production.pipelineRunning import SingleCorePipelineRunner
 from lsst.rubintv.production.podDefinition import PodDetails, PodFlavor
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import getAutomaticLocationConfig, getPodWorkerNumber
+from lsst.rubintv.production.utils import getAutomaticLocationConfig
 from lsst.summit.utils.utils import setupLogging
 
 setupLogging()
